Partially refactored acceptance tests
* Aggregation
* BoundedVarExpand
* ExpandInto
* Respected rel direction in CAPSPropertyGraphFactory
